# swagger/code-generation-api.yaml
openapi: 3.0.0
info:
  title: Code Generation API
  version: 1.0.0
  description: API for generating backend code files from schemas and APIs
paths:
...
  /code-generation/generate-backend:
    post:
      summary: Generate backend code files for a namespace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                namespaceId:
                  type: string
                schemas:
                  type: array
                  items:
                    type: object
                apis:
                  type: array
                  items:
                    type: object
                projectType:
                  type: string
                  enum: [nodejs, python]
                  default: nodejs
                namespaceName:
                  type: string
                  default: "Generated Project"
              required:
                - namespaceId
                - schemas
                - apis
      responses:
        '200':
          description: List of generated files
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                        name:
                          type: string
                        path:
                          type: string
        '400':
          description: Invalid input
        '500':
          description: Code generation failed
  /code-generation/files/{namespaceId}:
    get:
      summary: List generated files for a namespace
      parameters:
        - name: namespaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of files
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        path:
                          type: string
                        type:
                          type: string
                          enum: [file, directory]
                        size:
                          type: integer
        '500':
          description: Failed to list files
...